fix(tasks): skip unresolved mentions in task subscribers
Drop mentions that cannot be mapped to a project user id so comments-v2 global ids never land on the task. Extract the comments activity wrapper so FormEdit does not own the v1/v2 provider fork.
